+++ repyability/tests/test_rbd_importances.py
@@ -1,5 +1,5 @@
 """
-Tests many of the RBD class methods.
+Tests RBD's importance methods.
 
 Uses pytest fixtures located in conftest.py in the tests/ directory.
 """
@@ -7,257 +7,6 @@ import pytest
 
 from repyability.rbd.rbd import RBD
 
-
-# Check components are correct lengths
-def test_rbd_components(rbd1: RBD, rbd2: RBD):
-    assert len(rbd1.components) == 5
-    assert len(rbd2.components) == 8
-
-
-# Test get_all_path_sets()
-def test_rbd_get_all_path_sets(rbd1: RBD, rbd2: RBD):
-    assert list(rbd1.get_all_path_sets()) == [
-        ["source", "pump1", "valve", "sink"],
-        ["source", "pump2", "valve", "sink"],
-    ]
-    assert list(rbd2.get_all_path_sets()) == [
-        [1, 2, 3, 5, 6, 7, 8],
-        [1, 2, 4, 7, 8],
-    ]
-
-
-# sf()
-
-# Test sf() w/ simple series RBD
-def test_rbd_sf_series(rbd_series: RBD):
-    t = 5
-    assert (
-        pytest.approx(
-            rbd_series.components[2].sf(t)
-            * rbd_series.components[3].sf(t)
-            * rbd_series.components[4].sf(t)
-        )
-        == rbd_series.sf(t)[0]
-    )
-
-
-# Test get_min_path_sets()
-
-
-def test_rbd_get_min_path_sets_rbd_series(rbd_series: RBD):
-    assert {(1, 2, 3, 4, 5)} == rbd_series.get_min_path_sets()
-
-
-def test_rbd_get_min_path_sets_rbd_parallel(rbd_parallel: RBD):
-    assert {
-        (1, 2, 5),
-        (1, 3, 5),
-        (1, 4, 5),
-    } == rbd_parallel.get_min_path_sets()
-
-
-def test_rbd_get_min_path_sets_rbd1(rbd1: RBD):
-    assert {
-        ("source", "pump1", "valve", "sink"),
-        ("source", "pump2", "valve", "sink"),
-    } == rbd1.get_min_path_sets()
-
-
-def test_rbd_get_min_path_sets_rbd1_exclude_in_out_node(rbd1: RBD):
-    assert {
-        ("pump1", "valve"),
-        ("pump2", "valve"),
-    } == rbd1.get_min_path_sets(include_in_out_nodes=False)
-
-
-def test_rbd_get_min_path_sets_rbd2(rbd2: RBD):
-    assert {(1, 2, 3, 5, 6, 7, 8), (1, 2, 4, 7, 8)} == rbd2.get_min_path_sets()
-
-
-def test_rbd_get_min_path_sets_rbd3(rbd3: RBD):
-    assert {
-        (0, 1, 2, 6),
-        (0, 1, 5, 4, 6),
-        (0, 3, 4, 6),
-        (0, 3, 5, 2, 6),
-    } == rbd3.get_min_path_sets()
-
-
-# Test get_min_cut_sets()
-
-
-def test_rbd_get_min_cut_sets_rbd_series(rbd_series: RBD):
-    assert {
-        frozenset([2]),
-        frozenset([3]),
-        frozenset([4]),
-    } == rbd_series.get_min_cut_sets()
-
-
-def test_rbd_get_min_cut_sets_rbd_parallel(rbd_parallel: RBD):
-    assert {frozenset([2, 3, 4])} == rbd_parallel.get_min_cut_sets()
-
-
-def test_rbd_get_min_cut_sets_rbd1(rbd1: RBD):
-    assert {
-        frozenset(["pump1", "pump2"]),
-        frozenset(["valve"]),
-    } == rbd1.get_min_cut_sets()
-
-
-def test_rbd_get_min_cut_sets_rbd2(rbd2: RBD):
-    assert {
-        frozenset([2]),
-        frozenset([3, 4]),
-        frozenset([4, 5]),
-        frozenset([4, 6]),
-        frozenset([7]),
-    } == rbd2.get_min_cut_sets()
-
-
-def test_rbd_get_min_cut_sets_rbd3(rbd3: RBD):
-    assert {
-        frozenset([1, 3]),
-        frozenset([2, 4]),
-        frozenset([1, 4, 5]),
-        frozenset([2, 3, 5]),
-    } == rbd3.get_min_cut_sets()
-
-
-# Test sf() w/ simple series RBD
-def test_rbd_sf_parallel(rbd_parallel: RBD):
-    t = 2
-    assert (
-        pytest.approx(
-            (1 - rbd_parallel.components[2].sf(t))
-            * (1 - rbd_parallel.components[3].sf(t))
-            * (1 - rbd_parallel.components[4].sf(t))
-        )
-        == 1 - rbd_parallel.sf(t)[0]
-    )
-
-
-# Test sf() w/ a simple RBD with both parallel and series components
-def test_rbd_sf_composite(rbd1: RBD):
-    """Tests with an RBD with both parallel and series components."""
-    t = 2
-    assert (
-        pytest.approx(
-            (
-                1
-                - (1 - rbd1.components["pump1"].sf(t))
-                * (1 - rbd1.components["pump2"].sf(t))
-            )
-            * rbd1.components["valve"].sf(t)
-        )
-        == rbd1.sf(t)[0]
-    )
-
-
-# Test sf() w/ an RBD that cannot be reduced to parallel or series
-def test_rbd_sf_complex(rbd3: RBD):
-    assert pytest.approx(0.994780625) == rbd3.sf(1000)[0]
-
-
-# Test sf() w/ repeated component
-# i.e. Two nodes correspond to one node
-def test_rbd_sf_repeated_component(rbd_repeated_component_parallel: RBD):
-    t = 2
-    assert (
-        pytest.approx(
-            (1 - rbd_repeated_component_parallel.components[2].sf(t))
-            * (1 - rbd_repeated_component_parallel.components[3].sf(t))
-            * (1 - rbd_repeated_component_parallel.components[4].sf(t))
-            # Not 'component 5' as nodes 2 and 5 are both component 2
-        )
-        == 1 - rbd_repeated_component_parallel.sf(t)[0]
-    )
-
-
-# Test sf() w/ broken node
-def test_rbd_sf_broken_node(rbd_parallel: RBD):
-    t = 2
-    assert (
-        pytest.approx(
-            (1 - rbd_parallel.components[2].sf(t))
-            * (1 - rbd_parallel.components[4].sf(t))
-        )
-        == 1 - rbd_parallel.sf(t, broken_nodes=[3])[0]
-    )
-
-
-# Test sf() w/ broken component
-def test_rbd_sf_broken_component(rbd_repeated_component_parallel: RBD):
-    t = 2
-    assert (
-        pytest.approx(
-            (1 - rbd_repeated_component_parallel.components[2].sf(t))
-            * (1 - rbd_repeated_component_parallel.components[4].sf(t))
-        )
-        == 1 - rbd_repeated_component_parallel.sf(t, broken_components=[3])[0]
-    )
-    assert (
-        pytest.approx(
-            (1 - rbd_repeated_component_parallel.components[3].sf(t))
-            * (1 - rbd_repeated_component_parallel.components[4].sf(t))
-        )
-        == 1 - rbd_repeated_component_parallel.sf(t, broken_components=[2])[0]
-    )
-
-
-# Test sf() w/ working node
-def test_rbd_sf_working_node(rbd_series: RBD):
-    t = 2
-    assert (
-        pytest.approx(
-            rbd_series.components[2].sf(t) * rbd_series.components[4].sf(t)
-        )
-        == rbd_series.sf(t, working_nodes=[3])[0]
-    )
-
-
-# Test sf() w/ working component
-def test_rbd_sf_working_component(rbd_repeated_component_parallel: RBD):
-    t = 2
-    assert (
-        pytest.approx(1)
-        == rbd_repeated_component_parallel.sf(t, working_components=[2])[0]
-    )
-    assert (
-        pytest.approx(1)
-        == rbd_repeated_component_parallel.sf(t, working_components=[3])[0]
-    )
-
-
-# Test sf() w/ working node with repeated component
-def test_rbd_sf_working_node_repeated_component(
-    rbd_repeated_component_composite: RBD,
-):
-    rbd = rbd_repeated_component_composite
-    t = 2
-    assert (
-        pytest.approx(1 - rbd.components[2].ff(t) * rbd.components[3].ff(t))
-        == rbd.sf(t, working_nodes=[3])[0]
-    )
-    print(rbd.components[2].sf(t))
-
-
-# ff()
-
-# Test ff(), can just test parallel as ff() just calls 1 - sf()
-def test_rbd_ff(rbd_parallel: RBD):
-    t = 5
-    assert (
-        pytest.approx(
-            (1 - rbd_parallel.components[2].sf(t))
-            * (1 - rbd_parallel.components[3].sf(t))
-            * (1 - rbd_parallel.components[4].sf(t))
-        )
-        == rbd_parallel.ff(t)[0]
-    )
-
-
-# Importance calcs
 
 # Test birnbaum_importance() w/ composite RBD
 def test_rbd_birnbaum_importance(rbd1: RBD):
